+++ api-connector-service/src/main/java/com/mitsudoku/apiconnector/client/MovieDbClient.java
@@ -5,7 +5,6 @@ import com.mitsudoku.model.movie.MovieCreditsDto;
 import com.mitsudoku.model.movie.MovieDto;
 import com.mitsudoku.model.movie.ResponseDto;
 import com.mitsudoku.config.FeignConfig;
-import feign.Headers;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -24,6 +23,6 @@ public interface MovieDbClient {
     @RequestMapping(method = RequestMethod.GET, value = "movie/{movie-id}/credits")
 //    @Headers({"accept: application/json", "Authorization: Bearer ${movie.api.token}"})
     @Cacheable("movie-credits-cache")
-    MovieCreditsDto getMovieCredits(@PathVariable("movie-id") int id);
+    MovieCreditsDto getMovieCredits(@PathVariable("movie-id") Long id);
 
 }
